move StatementBuilder.ts to services package
